Fix wrong type assertion

+++ can.sensor.humiture.go
@@ -33,7 +33,7 @@ func (i *InstructionCANSensorHumiture) Execute(args ...string) (resp interface{}
 	if err != nil {
 		return resp, err
 	}
-	humiture, ok := resp.([]byte)
+	humiture, ok := resp.([]float64)
 	if !ok {
 		return resp, fmt.Errorf("invalid humiture response %#v", resp)
 	}
